made some changes

+++ app/cart.py
@@ -12,26 +12,29 @@ table = PrettyTable(['ID', 'Name', 'Price', 'qty'])
 
 def addToCart(params):
     if os.path.exists(__cart_db_location__):
-        item = get_item_by_path(f"{__item_db_location__}/{params[0]}.db")
-        item["qty"] = params[1]
-        _data_ = {
-            "items": [item]
-        }
-        if os.path.getsize(__cart_db_location__) == 0:
-            with open(__cart_db_location__, "w") as cart_file:
-                json.dump(_data_, cart_file)
-            print("Item added to the cart")
-        else:  
-            with open(__cart_db_location__, "r") as cart_file:
-                _cart_ = json.load(cart_file)
-                item_arr = _cart_["items"]
-                item_arr.append(item)
-                _updated_data_ = {
-                    "items": item_arr
-                }
+        if os.path.isfile(f"{__item_db_location__}/{params[0]}.db"):
+            item = get_item_by_path(f"{__item_db_location__}/{params[0]}.db")
+            item["qty"] = params[1]
+            _data_ = {
+                "items": [item]
+            }
+            if os.path.getsize(__cart_db_location__) == 0:
                 with open(__cart_db_location__, "w") as cart_file:
-                    json.dump(_updated_data_, cart_file)
+                    json.dump(_data_, cart_file)
                 print("Item added to the cart")
+            else:  
+                with open(__cart_db_location__, "r") as cart_file:
+                    _cart_ = json.load(cart_file)
+                    item_arr = _cart_["items"]
+                    item_arr.append(item)
+                    _updated_data_ = {
+                        "items": item_arr
+                    }
+                    with open(__cart_db_location__, "w") as cart_file:
+                        json.dump(_updated_data_, cart_file)
+                    print("Item added to the cart")
+        else:
+            print("No item for the given ID")
     else:
         item = get_item_by_path(f"{__item_db_location__}/{params[0]}.db")
         item["qty"] = params[1]
@@ -44,12 +47,15 @@ def addToCart(params):
 
 def viewCart():
     if os.path.exists(__cart_db_location__):
-        with open(__cart_db_location__, "r") as cart_file:
-            _cart_ = json.load(cart_file)
-            items = _cart_['items']   
-        for i in items:
-            table.add_row([i['id'], i['name'], i['sellingPrice'], i['qty']])
-        print(table)
+        if os.path.getsize(__cart_db_location__) != 0:
+            with open(__cart_db_location__, "r") as cart_file:
+                _cart_ = json.load(cart_file)
+                items = _cart_['items']   
+            for i in items:
+                table.add_row([i['id'], i['name'], i['sellingPrice'], i['qty']])
+            print(table)
+        else:
+            print("Cart is empty")
 
 def emptyCart():
     open(__cart_db_location__, 'w').close()
